NXP-5120: Avoid NPE in FileManageActionsBean when deleting temporary file

+++ nuxeo-platform-webapp-core/src/main/java/org/nuxeo/ecm/webapp/filemanager/FileManageActionsBean.java
@@ -608,7 +608,9 @@ public class FileManageActionsBean extends InputController implements
             } finally {
                 org.nuxeo.common.utils.FileUtils.close(stream);
                 // the content of the temporary blob has been
-                fileUploadHolder.getTempFile().delete();
+                if(fileUploadHolder.getTempFile() != null) {
+                    fileUploadHolder.getTempFile().delete();
+                }
             }
         } else {
             facesMessages.add(FacesMessage.SEVERITY_ERROR,
